@@ -12,17 +12,17 @@
   "private": true,
   "dependencies": {
     "@angular-material-components/datetime-picker": "^16.0.1",
-    "@angular/animations": "^17.3.12",
+    "@angular/animations": "^18.2.1",
     "@angular/cdk": "^17.3.10",
-    "@angular/common": "^17.3.12",
-    "@angular/compiler": "^17.3.12",
-    "@angular/core": "^17.3.12",
-    "@angular/forms": "^17.3.12",
+    "@angular/common": "^18.2.1",
+    "@angular/compiler": "^18.2.1",
+    "@angular/core": "^18.2.1",
+    "@angular/forms": "^18.2.1",
     "@angular/material": "^17.3.10",
     "@angular/material-luxon-adapter": "^17.3.10",
-    "@angular/platform-browser": "^17.3.12",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/router": "^17.3.12",
+    "@angular/platform-browser": "^18.2.1",
+    "@angular/platform-browser-dynamic": "^18.2.1",
+    "@angular/router": "^18.2.1",
     "@danielmoncada/angular-datetime-picker": "^16.0.1",
     "@ngrx/store": "^16.2.0",
     "angular-imask": "^7.1.3",
@@ -39,10 +39,10 @@
     "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.8",
-    "@angular/cli": "^17.3.8",
-    "@angular/compiler-cli": "^17.3.12",
-    "@angular/localize": "^17.3.12",
+    "@angular-devkit/build-angular": "^18.2.1",
+    "@angular/cli": "^18.2.1",
+    "@angular/compiler-cli": "^18.2.1",
+    "@angular/localize": "^18.2.1",
     "@types/jasmine": "~4.3.5",
     "@types/luxon": "^3.3.1",
     "@types/node": "^20.5.6",
