@@ -4,7 +4,7 @@ import { LayoutModule } from '@angular/cdk/layout';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { registerLocaleData } from '@angular/common';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import localeId from "@angular/common/locales/id";
 import { MAT_DATE_LOCALE } from '@angular/material/core';
@@ -151,144 +151,138 @@ import { PublicationFormsConfigurationsOverviewFormComponent } from './component
 
 registerLocaleData(localeId, 'id');
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    LoginComponent,
-    PageBlankComponent,
-    PageComponentsComponent,
-    PageNotFoundComponent,
-    PageUnderConstructionComponent,
-    CoreComponent,
-    HeaderComponent,
-    SidebarComponent,
-    ContentComponent,
-    FooterComponent,
-    OverviewComponent,
-    PublicationComponent,
-    PublicationFormComponent,
-    PublicationFormGridSystemsDefaultComponent,
-    PublicationFormGridSystemsMaterialComponent,
-    PublicationFormGridSystemsBootstrapComponent,
-    PublicationFormGridSystemsTailwindComponent,
-    PublicationFormRecursiveComponent,
-    PublicationDetailComponent,
-    PublicationDetailGridSystemsDefaultComponent,
-    PublicationDetailGridSystemsMaterialComponent,
-    PublicationDetailGridSystemsBootstrapComponent,
-    PublicationDetailGridSystemsTailwindComponent,
-    PublicationDetailRecursiveComponent,
-    CustomModalPublicationRemoveConfirmComponent,
-    CustomModalPublicationSubmitConfirmComponent,
-    CustomModalPublicationExitFormConfirmComponent,
-    CustomDialogPublicationRemoveConfirmComponent,
-    CustomDialogPublicationSubmitConfirmComponent,
-    CustomDialogPublicationExitFormConfirmComponent,
-    GhostTableComponent,
-    LoaderComponent,
-    DatetimePickerComponent,
-    DatePickerComponent,
-    MonthPickerComponent,
-    YearPickerComponent,
-    TimePickerComponent,
-    DateTimePickerComponent,
-    OwlDatetimePickerComponent,
-    OwlDatePickerComponent,
-    OwlMonthPickerComponent,
-    OwlYearPickerComponent,
-    OwlTimePickerComponent,
-    OwlDateTimePickerComponent,
-    OwlDateRangePickerComponent,
-    OwlTimeRangePickerComponent,
-    OwlDateTimeRangePickerComponent,
-    DialogConfirmComponent,
-    DialogAlertComponent,
-    DialogPromptComponent,
-    ModalConfirmComponent,
-    ModalAlertComponent,
-    ModalPromptComponent,
-    DynamicFormComponent,
-    DFFieldTextComponent,
-    DFFieldNumberComponent,
-    DFFieldTextareaComponent,
-    DFFieldDateComponent,
-    DFFieldTimeComponent,
-    DFFieldDatetimeComponent,
-    DFFieldYearComponent,
-    DFFieldFileComponent,
-    DFFieldImageComponent,
-    DFFieldFileUploadComponent,
-    DFFieldImageUploadComponent,
-    DFFieldColorComponent,
-    DFFieldCheckboxComponent,
-    DFFieldRadioComponent,
-    DFFieldSliderComponent,
-    DFFieldSelectComponent,
-    DFFieldNGXSelectComponent,
-    DfFieldEmailComponent,
-    DfFieldTelComponent,
-    DfFieldUrlComponent,
-    DfFieldPasswordComponent,
-    DfFieldSearchComponent,
-    DFFieldMonthComponent,
-    DFFieldOwlDatetimeComponent,
-    DFFieldOwlDateComponent,
-    DFFieldOwlDateTimeComponent,
-    DFFieldOwlTimeComponent,
-    DFFieldOwlMonthComponent,
-    DFFieldOwlYearComponent,
-    DFFieldOwlDateRangeComponent,
-    DFFieldOwlTimeRangeComponent,
-    DFFieldOwlDateTimeRangeComponent,
-    DFFieldMultipleComponent,
-    DFAlertComponent,
-    DFDialogFileUploadPromptComponent,
-    DFWrapperStepperComponent,
-    DFWrapperAccordionComponent,
-    DFWrapperPanelComponent,
-    DFWrapperMultipleComponent,
-    DFCustomAlertInfoNoDataComponent,
-    ConfigurationsComponent,
-    PublicationFormsConfigurationsComponent,
-    PublicationTypesMasterDataConfigurationsComponent,
-    PublicationGeneralTypesMasterDataConfigurationsComponent,
-    PublicationFormVersionsConfigurationsComponent,
-    PublicationFormsConfigurationsFormComponent,
-    PublicationFormsVersionsConfigurationsFormComponent,
-    PublicationFormsVersionsConfigurationsDetailComponent,
-    PublicationFormsConfigurationsDetailComponent,
-    PublicationFormsInitialConfigurationsFormComponent,
-    PublicationFormsGeneralConfigurationsFormComponent,
-    PublicationFormsAdvancedConfigurationsFormComponent,
-    PublicationFormsConfigurationsOverviewFormComponent,
-  ],
-  imports: [
-    BrowserModule,
-    LayoutModule,
-    AppRoutingModule,
-    BrowserAnimationsModule,
-    CoreModule,
-    FormsModule,
-    ReactiveFormsModule,
-    MaterialModule,
-    BootstrapModule,
-    HttpClientModule,
-    LuxonModule,
-    NgxSelectModule,
-    NgxMatDatetimePickerModule,
-    NgxMatTimepickerModule,
-    NgxMatNativeDateModule,
-    OwlDateTimeModule,
-    OwlNativeDateTimeModule,
-    NgxDocViewerModule,
-  ],
-  providers: [
-    { provide: LOCALE_ID, useValue: 'id-ID' },
-    { provide: MAT_DATE_LOCALE, useValue: 'id-ID' },
-    { provide: OWL_DATE_TIME_LOCALE, useValue: 'id-ID' },
-    { provide: ENV, useValue: CONFIG },
-    AppService, AppGeneralService
-  ],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        AppComponent,
+        LoginComponent,
+        PageBlankComponent,
+        PageComponentsComponent,
+        PageNotFoundComponent,
+        PageUnderConstructionComponent,
+        CoreComponent,
+        HeaderComponent,
+        SidebarComponent,
+        ContentComponent,
+        FooterComponent,
+        OverviewComponent,
+        PublicationComponent,
+        PublicationFormComponent,
+        PublicationFormGridSystemsDefaultComponent,
+        PublicationFormGridSystemsMaterialComponent,
+        PublicationFormGridSystemsBootstrapComponent,
+        PublicationFormGridSystemsTailwindComponent,
+        PublicationFormRecursiveComponent,
+        PublicationDetailComponent,
+        PublicationDetailGridSystemsDefaultComponent,
+        PublicationDetailGridSystemsMaterialComponent,
+        PublicationDetailGridSystemsBootstrapComponent,
+        PublicationDetailGridSystemsTailwindComponent,
+        PublicationDetailRecursiveComponent,
+        CustomModalPublicationRemoveConfirmComponent,
+        CustomModalPublicationSubmitConfirmComponent,
+        CustomModalPublicationExitFormConfirmComponent,
+        CustomDialogPublicationRemoveConfirmComponent,
+        CustomDialogPublicationSubmitConfirmComponent,
+        CustomDialogPublicationExitFormConfirmComponent,
+        GhostTableComponent,
+        LoaderComponent,
+        DatetimePickerComponent,
+        DatePickerComponent,
+        MonthPickerComponent,
+        YearPickerComponent,
+        TimePickerComponent,
+        DateTimePickerComponent,
+        OwlDatetimePickerComponent,
+        OwlDatePickerComponent,
+        OwlMonthPickerComponent,
+        OwlYearPickerComponent,
+        OwlTimePickerComponent,
+        OwlDateTimePickerComponent,
+        OwlDateRangePickerComponent,
+        OwlTimeRangePickerComponent,
+        OwlDateTimeRangePickerComponent,
+        DialogConfirmComponent,
+        DialogAlertComponent,
+        DialogPromptComponent,
+        ModalConfirmComponent,
+        ModalAlertComponent,
+        ModalPromptComponent,
+        DynamicFormComponent,
+        DFFieldTextComponent,
+        DFFieldNumberComponent,
+        DFFieldTextareaComponent,
+        DFFieldDateComponent,
+        DFFieldTimeComponent,
+        DFFieldDatetimeComponent,
+        DFFieldYearComponent,
+        DFFieldFileComponent,
+        DFFieldImageComponent,
+        DFFieldFileUploadComponent,
+        DFFieldImageUploadComponent,
+        DFFieldColorComponent,
+        DFFieldCheckboxComponent,
+        DFFieldRadioComponent,
+        DFFieldSliderComponent,
+        DFFieldSelectComponent,
+        DFFieldNGXSelectComponent,
+        DfFieldEmailComponent,
+        DfFieldTelComponent,
+        DfFieldUrlComponent,
+        DfFieldPasswordComponent,
+        DfFieldSearchComponent,
+        DFFieldMonthComponent,
+        DFFieldOwlDatetimeComponent,
+        DFFieldOwlDateComponent,
+        DFFieldOwlDateTimeComponent,
+        DFFieldOwlTimeComponent,
+        DFFieldOwlMonthComponent,
+        DFFieldOwlYearComponent,
+        DFFieldOwlDateRangeComponent,
+        DFFieldOwlTimeRangeComponent,
+        DFFieldOwlDateTimeRangeComponent,
+        DFFieldMultipleComponent,
+        DFAlertComponent,
+        DFDialogFileUploadPromptComponent,
+        DFWrapperStepperComponent,
+        DFWrapperAccordionComponent,
+        DFWrapperPanelComponent,
+        DFWrapperMultipleComponent,
+        DFCustomAlertInfoNoDataComponent,
+        ConfigurationsComponent,
+        PublicationFormsConfigurationsComponent,
+        PublicationTypesMasterDataConfigurationsComponent,
+        PublicationGeneralTypesMasterDataConfigurationsComponent,
+        PublicationFormVersionsConfigurationsComponent,
+        PublicationFormsConfigurationsFormComponent,
+        PublicationFormsVersionsConfigurationsFormComponent,
+        PublicationFormsVersionsConfigurationsDetailComponent,
+        PublicationFormsConfigurationsDetailComponent,
+        PublicationFormsInitialConfigurationsFormComponent,
+        PublicationFormsGeneralConfigurationsFormComponent,
+        PublicationFormsAdvancedConfigurationsFormComponent,
+        PublicationFormsConfigurationsOverviewFormComponent,
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        LayoutModule,
+        AppRoutingModule,
+        BrowserAnimationsModule,
+        CoreModule,
+        FormsModule,
+        ReactiveFormsModule,
+        MaterialModule,
+        BootstrapModule,
+        LuxonModule,
+        NgxSelectModule,
+        NgxMatDatetimePickerModule,
+        NgxMatTimepickerModule,
+        NgxMatNativeDateModule,
+        OwlDateTimeModule,
+        OwlNativeDateTimeModule,
+        NgxDocViewerModule], providers: [
+        { provide: LOCALE_ID, useValue: 'id-ID' },
+        { provide: MAT_DATE_LOCALE, useValue: 'id-ID' },
+        { provide: OWL_DATE_TIME_LOCALE, useValue: 'id-ID' },
+        { provide: ENV, useValue: CONFIG },
+        AppService, AppGeneralService,
+        provideHttpClient(withInterceptorsFromDi())
+    ] })
 export class AppModule { }
